WG2D

+++ weatherGenerator2D/wg2D.h
@@ -91,7 +91,7 @@ private:
     void precipitationP00P10();
     void precipitationCorrelationMatrices();
     void precipitationMultisiteOccurrenceGeneration();
-    void spatialIterationOccurrence(double ** M, float **K, float** occurrences, double** matrixOccurrence, double** normalizedMatrixRandom, float **transitionNormal, int lengthSeries);
+    void spatialIterationOccurrence(double ** M, double **K, double **occurrences, double** matrixOccurrence, double** normalizedMatrixRandom, double **transitionNormal, int lengthSeries);
     void temperatureCompute();
